refactor: styles structure
