--- src/consts/routerPaths.ts
@@ -1,6 +1,0 @@
-export const paths = {
-  welcome: '/',
-  main: '/graphiql',
-  signIn: '/auth',
-  signUp: '/auth?register=true',
-};
