+++ src/enums/routerPaths.ts
@@ -1,0 +1,6 @@
+export enum paths {
+  welcome = '/',
+  main = '/graphiql',
+  signIn = '/auth',
+  signUp = '/auth?register=true',
+}
