+++ src/pages/auth.tsx
@@ -7,7 +7,7 @@ import { auth } from '@/services/authService';
 import styles from '@/styles/Auth.module.scss';
 import Image from 'next/image';
 import playgroundLogo from '@/assets/images/playgroundLogo.png';
-import { paths } from '@/consts/routerPaths';
+import { paths } from '@/enums/routerPaths';
 
 const Auth = () => {
   const {
