+++ src/styles/Auth.module.scss
@@ -7,18 +7,20 @@
   @include elements-style.center-content;
   background-color: colors.$white;
   padding: 30px 0;
+
+  &__container {
+    max-width: 500px;
+    flex: 1 1 auto;
+    @include elements-style.flex-column;
+    text-align: center;
+  }
+
+  &__logo {
+    width: 100%;
+    height: auto;
+    object-fit: cover;
+    object-position: center;
+  }
 }
 
-.auth__container {
-  max-width: 500px;
-  flex: 1 1 auto;
-  @include elements-style.flex-column;
-  text-align: center;
-}
 
-.auth__logo {
-  width: 100%;
-  height: auto;
-  object-fit: cover;
-  object-position: center;
-}
